Move bot_v2 to bot
