+++ bot/cogs/randoms.py
