--- bot/methods/utils.py
@@ -1,72 +1,0 @@
-import json
-
-from cache import cache
-
-
-def change_prefix(params=None, data=None):
-    if not params:
-        return {"error": "USAGE"}
-    prefix = params.split()[0]
-    message = {}
-    message["message"] = f"Prefix updated to {prefix}"
-    data["general"]["prefix"] = prefix
-    message["data"] = data
-    return message
-
-
-def get_usage(params=None, data=None):
-    if not params:
-        return {"error": "USAGE"}
-
-    usage_command = params.split()[0].lower()
-
-    commands = cache.get("commands")
-    if not commands:
-        with open("commands.json") as f:
-            commands = json.load(f)["commands"]
-        cache.add("commands", commands, 1440)
-
-    found_command = {}
-    for command in commands:
-        if usage_command == command["command"]:
-            found_command = command
-
-    if not found_command:
-        return {
-            "error": "USAGE",
-            "message": f'Command not recognized. Use "{data["general"]["prefix"]}list" to see a full list of commands.',
-        }
-
-    message = {}
-    usages = (
-        found_command["usage"]
-        if type(found_command["usage"]) is list
-        else [found_command["usage"]]
-    )
-    message["message"] = f"Usage:"
-    for usage in usages:
-        message[
-            "message"
-        ] += f'\n`{data["general"]["prefix"]}{found_command["command"]} {usage}`'
-    return message
-
-
-def list_commands(params=None, data=None):
-    commands = cache.get("commands")
-    if not commands:
-        with open("commands.json") as f:
-            commands = json.load(f)["commands"]
-        cache.add("commands", commands, 1440)
-
-    message = {}
-    message["message"] = "List of commands and their descriptions:\n```"
-
-    for command in commands:
-        message[
-            "message"
-        ] += f'\n{data["general"]["prefix"]}{command["command"]} - ' + (
-            f'{command["description"]}' if "description" in command else ""
-        )
-    message["message"] += "\n```"
-
-    return message
