+++ bot/requirements.txt
@@ -1,15 +1,25 @@
 aiohttp==3.6.3
+appdirs==1.4.4
 async-timeout==3.0.1
 attrs==20.2.0
+black==20.8b1
 certifi==2020.6.20
 chardet==3.0.4
+click==7.1.2
 discord==1.0.1
+discord-py-slash-command==1.1.2
 discord.py==1.5.1
 idna==2.10
 multidict==4.7.6
+mypy-extensions==0.4.3
 numpy==1.19.4
+pathspec==0.8.1
 Pillow==8.0.1
 pkg-resources==0.0.0
+regex==2021.4.4
 requests==2.24.0
+toml==0.10.2
+typed-ast==1.4.3
+typing-extensions==3.7.4.3
 urllib3==1.25.11
 yarl==1.5.1
