+++ bot/resources/.gitignore
