--- bot/resources/statuses.txt
@@ -1,12 +1,0 @@
-a movie
-cat videos
-my cat
-my dog
-my grandkids
-my watches
-my weight
-netflix
-television
-the office, again
-you :)
-youtube
