--- bot_v2/data/guilds/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
