--- bot_v2/data/users/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
