--- bot_v2/logs/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
