--- bot_v2/setup
@@ -1,8 +1,0 @@
-#!/bin/bash
-python3 -m venv .venv
-source .venv/bin/activate
-
-pip3 install --upgrade pip
-pip3 install --upgrade wheel
-
-pip3 install -r ./requirements.txt
