--- bot_v2/ve
@@ -1,1 +1,0 @@
-source .venv/bin/activate
